move restore-related exceptions to the phone module
